requirements.txt
Flask==1.1.1
Jinja2==2.11.1
google-cloud-firestore==1.6.2